+++ common/src/main/java/immersive_aircraft/DataLoaders.java
@@ -1,6 +1,7 @@
 package immersive_aircraft;
 
 import immersive_aircraft.cobalt.registration.Registration;
+import immersive_aircraft.data.AircraftDataLoader;
 import immersive_aircraft.data.BaseStatDataLoader;
 import immersive_aircraft.data.UpgradeDataLoader;
 import immersive_aircraft.resources.BBModelLoader;
@@ -14,6 +15,7 @@ public class DataLoaders {
     static {
         Registration.registerDataLoader("aircraft_upgrades", new UpgradeDataLoader());
         Registration.registerDataLoader("aircraft_stats", new BaseStatDataLoader());
+        Registration.registerDataLoader("aircraft", new AircraftDataLoader());
 
         Registration.registerResourceLoader("objects_obj", new ObjectLoader());
         Registration.registerResourceLoader("objects_bbmodel", new BBModelLoader());