+++ common/src/main/java/immersive_aircraft/Messages.java
@@ -14,6 +14,7 @@ public class Messages {
         NetworkHandler.registerMessage(CollisionMessage.class, CollisionMessage::new);
         NetworkHandler.registerMessage(AircraftUpgradesMessage.class, AircraftUpgradesMessage::new);
         NetworkHandler.registerMessage(AircraftBaseUpgradesMessage.class, AircraftBaseUpgradesMessage::new);
+        NetworkHandler.registerMessage(AircraftDataMessage.class, AircraftDataMessage::new);
         NetworkHandler.registerMessage(FireMessage.class, FireMessage::new);
         NetworkHandler.registerMessage(FireResponse.class, FireResponse::new);
     }