+++ common/src/main/java/immersive_aircraft/entity/AirplaneEntity.java
@@ -1,6 +1,6 @@
 package immersive_aircraft.entity;
 
-import immersive_aircraft.entity.misc.AircraftProperties;
+import immersive_aircraft.item.upgrade.AircraftStat;
 import net.minecraft.world.entity.EntityType;
 import net.minecraft.world.level.Level;
 import net.minecraft.world.phys.Vec3;
@@ -12,18 +12,6 @@ public abstract class AirplaneEntity extends EngineAircraft {
     // todo property
     private static final float PUSH_SPEED = 0.025f;
 
-    private final AircraftProperties properties = new AircraftProperties(this)
-            .setYawSpeed(5.0f)
-            .setPitchSpeed(4.0f)
-            .setEngineSpeed(0.0225f)
-            .setGlideFactor(0.05f)
-            .setDriftDrag(0.01f)
-            .setLift(0.15f)
-            .setRollFactor(45.0f)
-            .setGroundPitch(4.0f)
-            .setWindSensitivity(0.025f)
-            .setMass(1.0f);
-
     public AirplaneEntity(EntityType<? extends AircraftEntity> entityType, Level world, boolean canExplodeOnCrash) {
         super(entityType, world, canExplodeOnCrash);
     }
@@ -31,11 +19,6 @@ public abstract class AirplaneEntity extends EngineAircraft {
     @Override
     protected boolean useAirplaneControls() {
         return true;
-    }
-
-    @Override
-    public AircraftProperties getProperties() {
-        return properties;
     }
 
     @Override
@@ -74,7 +57,7 @@ public abstract class AirplaneEntity extends EngineAircraft {
         Vec3 direction = getForwardDirection();
 
         // speed
-        float thrust = (float) (Math.pow(getEnginePower(), 2.0) * properties.getEngineSpeed());
+        float thrust = (float) (Math.pow(getEnginePower(), 2.0) * getProperties().get(AircraftStat.ENGINE_SPEED));
         if (onGround && getEngineTarget() < 1.0) {
             thrust = PUSH_SPEED / (1.0f + (float) getDeltaMovement().length() * 5.0f) * pressingInterpolatedZ.getSmooth() * (1.0f - getEnginePower());
         }