+++ common/src/main/java/immersive_aircraft/entity/AirshipEntity.java
@@ -5,9 +5,8 @@ import com.mojang.math.Vector3f;
 import com.mojang.math.Vector4f;
 import immersive_aircraft.Items;
 import immersive_aircraft.Sounds;
-import immersive_aircraft.entity.misc.AircraftProperties;
 import immersive_aircraft.entity.misc.Trail;
-import immersive_aircraft.entity.misc.VehicleInventoryDescription;
+import immersive_aircraft.item.upgrade.AircraftStat;
 import net.minecraft.core.particles.ParticleTypes;
 import net.minecraft.sounds.SoundEvent;
 import net.minecraft.world.entity.EntityType;
@@ -18,33 +17,6 @@ import net.minecraft.world.phys.Vec3;
 import java.util.List;
 
 public class AirshipEntity extends Rotorcraft {
-    private final AircraftProperties properties = new AircraftProperties(this)
-            .setYawSpeed(5.0f)
-            .setEngineSpeed(0.02f)
-            .setVerticalSpeed(0.025f)
-            .setGlideFactor(0.0f)
-            .setDriftDrag(0.01f)
-            .setLift(0.1f)
-            .setRollFactor(5.0f)
-            .setWindSensitivity(0.05f)
-            .setMass(3.0f);
-
-    private static final VehicleInventoryDescription inventoryDescription = new VehicleInventoryDescription()
-            .addSlot(VehicleInventoryDescription.SlotType.WEAPON, 8 + 18 * 2 + 6, 8 + 6)
-            .addSlot(VehicleInventoryDescription.SlotType.WEAPON, 8 + 18 * 2 + 28, 8 + 6)
-            .addSlot(VehicleInventoryDescription.SlotType.UPGRADE, 8 + 18 * 2 + 6, 8 + 6 + 22)
-            .addSlot(VehicleInventoryDescription.SlotType.UPGRADE, 8 + 18 * 2 + 28, 8 + 6 + 22)
-            .addSlot(VehicleInventoryDescription.SlotType.BANNER, 8 + 18 * 2 + 6, 8 + 6 + 22 * 2)
-            .addSlot(VehicleInventoryDescription.SlotType.DYE, 8 + 18 * 2 + 28, 8 + 6 + 22 * 2)
-            .addSlot(VehicleInventoryDescription.SlotType.BOILER, 8 + 9, 8 + 36)
-            .addSlots(VehicleInventoryDescription.SlotType.INVENTORY, 8 + 18 * 5, 8, 4, 4)
-            .build();
-
-    @Override
-    public VehicleInventoryDescription getInventoryDescription() {
-        return inventoryDescription;
-    }
-
     public AirshipEntity(EntityType<? extends AircraftEntity> entityType, Level world) {
         super(entityType, world, true);
     }
@@ -56,11 +28,6 @@ public class AirshipEntity extends Rotorcraft {
 
     protected SoundEvent getEngineSound() {
         return Sounds.PROPELLER_SMALL.get();
-    }
-
-    @Override
-    public AircraftProperties getProperties() {
-        return properties;
     }
 
     @Override
@@ -88,16 +55,6 @@ public class AirshipEntity extends Rotorcraft {
         return Items.AIRSHIP.get();
     }
 
-    final List<List<Vec3>> PASSENGER_POSITIONS = List.of(
-            List.of(
-                    new Vec3(0.0f, -0.1f, 0.0f)
-            ),
-            List.of(
-                    new Vec3(0.0f, -0.1f, 0.4f),
-                    new Vec3(0.0f, -0.1f, -0.3f)
-            )
-    );
-
     private final List<Trail> trails = List.of(new Trail(15, 0.5f));
 
     public List<Trail> getTrails() {
@@ -116,10 +73,6 @@ public class AirshipEntity extends Rotorcraft {
         getTrails().get(index).add(p0, p1, trailStrength);
     }
 
-    protected List<List<Vec3>> getPassengerPositions() {
-        return PASSENGER_POSITIONS;
-    }
-
     @Override
     protected float getGravity() {
         return wasTouchingWater ? 0.04f : (1.0f - getEnginePower()) * super.getGravity();
@@ -132,13 +85,13 @@ public class AirshipEntity extends Rotorcraft {
         setEngineTarget(1.0f);
 
         // up and down
-        setDeltaMovement(getDeltaMovement().add(0.0f, getEnginePower() * properties.getVerticalSpeed() * pressingInterpolatedY.getSmooth(), 0.0f));
+        setDeltaMovement(getDeltaMovement().add(0.0f, getEnginePower() * getProperties().get(AircraftStat.VERTICAL_SPEED) * pressingInterpolatedY.getSmooth(), 0.0f));
 
         // get pointing direction
         Vec3 direction = getForwardDirection();
 
         // accelerate
-        float thrust = (float) (Math.pow(getEnginePower(), 5.0) * properties.getEngineSpeed()) * pressingInterpolatedZ.getSmooth();
+        float thrust = (float) (Math.pow(getEnginePower(), 5.0) * getProperties().get(AircraftStat.ENGINE_SPEED)) * pressingInterpolatedZ.getSmooth();
         setDeltaMovement(getDeltaMovement().add(direction.scale(thrust)));
     }