+++ common/src/main/java/immersive_aircraft/entity/Rotorcraft.java
@@ -1,5 +1,6 @@
 package immersive_aircraft.entity;
 
+import immersive_aircraft.item.upgrade.AircraftStat;
 import net.minecraft.util.Mth;
 import net.minecraft.world.entity.EntityType;
 import net.minecraft.world.level.Level;
@@ -33,8 +34,8 @@ public abstract class Rotorcraft extends EngineAircraft {
         Vec3 velocity = getDeltaMovement().multiply(1.0f, 0.0f, 1.0f);
         double drag = Math.abs(direction.dot(velocity.normalize()));
         Vec3 newVelocity = velocity.normalize()
-                .lerp(direction, getProperties().getLift())
-                .scale(velocity.length() * (drag * getProperties().getDriftDrag() + (1.0 - getProperties().getDriftDrag())));
+                .lerp(direction, getProperties().get(AircraftStat.LIFT))
+                .scale(velocity.length() * (drag * getProperties().get(AircraftStat.FRICTION) + (1.0 - getProperties().get(AircraftStat.FRICTION))));
         setDeltaMovement(
                 newVelocity.x,
                 getDeltaMovement().y,