+++ common/src/main/java/immersive_aircraft/item/upgrade/AircraftUpgradeRegistry.java
@@ -4,21 +4,9 @@ import net.minecraft.world.item.Item;
 
 import java.util.HashMap;
 import java.util.Map;
-import java.util.Set;
 
 public class AircraftUpgradeRegistry {
-
     public static final AircraftUpgradeRegistry INSTANCE = new AircraftUpgradeRegistry();
-    public static final Map<String, AircraftStat> STATS = new HashMap<>();
-
-    static {
-        STATS.put("strength", AircraftStat.STRENGTH);
-        STATS.put("acceleration", AircraftStat.ACCELERATION);
-        STATS.put("durability", AircraftStat.DURABILITY);
-        STATS.put("wind", AircraftStat.WIND);
-        STATS.put("friction", AircraftStat.FRICTION);
-        STATS.put("fuel", AircraftStat.FUEL);
-    }
 
     private final Map<Item, AircraftUpgrade> itemUpgrades = new HashMap<>();
 
@@ -39,10 +27,6 @@ public class AircraftUpgradeRegistry {
         itemUpgrades.clear();
     }
 
-    public Set<Item> getKeys() {
-        return itemUpgrades.keySet();
-    }
-
     public Map<Item, AircraftUpgrade> getAll() {
         return itemUpgrades;
     }
@@ -50,5 +34,4 @@ public class AircraftUpgradeRegistry {
     public boolean hasUpgrade(Item item) {
         return itemUpgrades.containsKey(item);
     }
-
 }